Rework load animation and quiet mode

+++ towa-connect.sh
@@ -5,81 +5,59 @@ source config.sh
 towaScriptsSrc="${mainPath}/src/towa" 
 towaTestsSrc="${mainPath}/test/towa"
 
+echo -en "\e[36m" # Cyan
 cat splashscreen.txt
 
-rm -r towa
+echo -en "\e[90m" # grey
 
-echo "- creating output folder"
+[ -d "towa" ] &&  rm -r towa
+
 mkdir towa
 cp $towaScriptsSrc/*.java towa
 cp $towaTestsSrc/*.java towa
-echo "- output folder filled with scripts"
 
-echo
-
-echo "Connecting to delete actual files"
-
-
-#Revers tunnel
 ssh -R 2222:localhost:22 $iutuser@info-ssh1.iut.u-bordeaux.fr << EOF
   cd ~/iut-remise/towa/info_s1/$iutuser/depot
-  echo "---------------------------"
-  echo "removing sortie.log"
   rm sortie.log
-  echo "removing actual towa folder"
   rm -r towa
-  echo "---------------------------"
 EOF
-
-echo "Connecting to send new files"
 
 rsync -r ./towa $iutuser@info-ssh1.iut.u-bordeaux.fr:~/iut-remise/towa/info_s1/$iutuser/depot
 
 rm -r towa
 
-
-echo "Connecting to retrieve sortie.log"
-#Revers tunnel
 ssh -R 2222:localhost:22 $iutuser@info-ssh1.iut.u-bordeaux.fr << EOF
+  
   cd ~/iut-remise/towa/info_s1/$iutuser/depot
-  echo "---------------------------"
-
-  echo "waiting for sortie.log to get created"
+  
+  echo -en "\e[92mEn attente de sortie.log : \e[33m["
   while [ ! -f sortie.log ]
   do 
-    echo -n "."
-    sleep 0.2
+    echo -ne "▄"
+    sleep 0.5
+    echo -ne "▀"
+    sleep 0.5
   done
-
-  echo
-  echo "-----sortie.log created----"
-
-  echo "waiting for sortie.log to get filled"
-
-  # notfound=1
-
-  # while (( notfound )); do
-  
-  #   wc -l sortie.log | cut -f 1 -d " "
-
-  #   #if [[ $size -gt 0 ]]; then
-  #   #  notfound=0;
-  #   #fi 
-
-  #   # echo -n "/"
-    
-  #   sleep 0.2
-  # done
-
-  #SLEEP DE LA DEFAITE
-  sleep 20
-
-  echo
-  echo "-----sortie.log filled----"
+  echo "]"
 EOF
+
+echo
+
+counter=0
+
+echo -en "\e[92mContenu en attente       : \e[33m["
+while (( $counter < 20 ))
+do 
+  echo -ne "▄"
+  sleep 0.5
+  echo -ne "▀"
+  sleep 0.5
+  counter=$counter+1
+done
+echo "]"
 
 rsync -r $iutuser@info-ssh1.iut.u-bordeaux.fr:~/iut-remise/towa/info_s1/$iutuser/depot/sortie.log ./
 
+echo
+echo -en "\e[37m"
 cat sortie.log
-
-echo "COMPLETE ;)"
